renamed files under the epic tests directory to make them shorter and easier to read
