+++ public/src/main.tsx
@@ -10,7 +10,7 @@ import {
   LiveblogEpicTestsForm,
   AppleNewsEpicTestsForm,
   AMPEpicTestsForm,
-} from './components/channelManagement/epicTests/epicTestsForm';
+} from './components/channelManagement/epicTests/testsForm';
 import {
   BannerTestsForm1,
   BannerTestsForm2,
